+++ gpxity/__init__.py
@@ -18,7 +18,7 @@ from .auth import *
 
 __all__ = [
     'Track', 'Fences', 'Lifetrack', 'Locate', 'Directory', 'GPSIES', 'MMT', 'TrackMMT', 'Openrunner',
-    'Authenticate', 'ServerDirectory', 'BackendDiff', 'WPTrackserver', 'Mailer', 'VERSION']
+    'Authenticate', 'BackendDiff', 'WPTrackserver', 'Mailer', 'VERSION']
 
 
 def prepare_backends():