+++ gpxity/backends/__init__.py
@@ -6,7 +6,6 @@
 """define things that should be visible to the user."""
 
 from .directory import *
-from .server_directory import *
 from .mmt import *
 from .gpsies import *
 from .trackmmt import *